smart_house_udp_therm_thread 1. add global freeze

+++ otus/smart_home/smart_home/src/devices/device.rs
@@ -52,7 +52,7 @@ pub struct SmartSocket {
 }
 pub struct SmartThermometer {
     pub info: DeviceInfo,
-    pub temperature: u32,
+    pub temperature: f32,
 }
 
 pub trait Device {
